Plutus.Contract.Test.ContractModel.Interface

Contract models

class (Typeable state, Show state, Show (Action state), Eq (Action state), HasSymbolics (Action state), HasVariables (Action state), HasVariables state, Generic state, forall w s e p. Eq (ContractInstanceKey state w s e p), forall w s e p. Show (ContractInstanceKey state w s e p)) => ContractModel state where

A ContractModel instance captures everything that is needed to generate and run tests of a contract or set of contracts. It specifies among other things

what operations are supported by the contract (Action),

when they are valid (precondition),

how to generate random actions (arbitraryAction),

how the operations affect the state (nextState), and

how to run the operations in the emulator (perform)

Minimal complete definition

instanceWallet, arbitraryAction, initialState, nextState, instanceContract, perform

Associated Types

data Action state

The type of actions that are supported by the contract. An action usually represents a single callEndpoint or a transfer of tokens, but it can be anything that can be interpreted in the EmulatorTrace monad.

data ContractInstanceKey state :: * -> Row * -> * -> * -> *

To be able to call a contract endpoint from a wallet a ContractHandle is required. These are managed by the test framework and all the user needs to do is provide this contract instance key type representing the different contract instances that a test needs to work with, and when creating a property (see propRunActions_) provide a list of contract instance keys together with their wallets and contracts (a ContractInstanceSpec). Contract instance keys are indexed by the observable state, schema, and error type of the contract and should be defined as a GADT. For example, a handle type for a contract with one seller and multiple buyers could look like this.

 data ContractInstanceKey MyModel w s e where
     Buyer  :: Wallet -> ContractInstanceKey MyModel MyObsState MySchema MyError MyParams
     Seller :: ContractInstanceKey MyModel MyObsState MySchema MyError MyParams

Methods

instanceWallet :: ContractInstanceKey state w s e p -> Wallet

Get the wallet that the contract running at a specific ContractInstanceKey should run in

instanceTag :: forall w s e p. SchemaConstraints w s e => ContractInstanceKey state w s e p -> ContractInstanceTag

The ContractInstanceTag of an instance key for a wallet. Defaults to walletInstanceTag. You must override this if you have multiple instances per wallet.

arbitraryAction :: ModelState state -> Gen (Action state)

Given the current model state, provide a QuickCheck generator for a random next action. This is used in the Arbitrary instance for Actionss as well as by anyAction and anyActions.

actionName :: Action state -> String

The name of an Action, used to report statistics.

waitProbability :: ModelState state -> Double

The probability that we will generate a WaitUntil in a given state

arbitraryWaitInterval :: ModelState state -> Gen Slot

Control the distribution of how long WaitUntil waits

initialState :: state

The initial state, before any actions have been performed.

initialInstances :: [StartContract state]

The initial handles

precondition :: ModelState state -> Action state -> Bool

The precondition function decides if a given action is valid in a given state. Typically actions generated by arbitraryAction will satisfy the precondition, but if they don't they will be discarded and another action will be generated. More importantly, the preconditions are used when shrinking (see shrinkAction) to ensure that shrunk test cases still make sense.

If an explicit action in a DL scenario violates the precondition an error is raised.

nextReactiveState :: Slot -> Spec state ()

nextReactiveState is run every time the model waits for a slot to be reached. This can be used to model reactive components of off-chain code.

nextState :: Action state -> Spec state ()

This is where the model logic is defined. Given an action, nextState specifies the effects running that action has on the model state. It runs in the Spec monad, which is a state monad over the ModelState.

startInstances :: ModelState state -> Action state -> [StartContract state]

Start new contract instances

instanceContract :: (forall t. HasSymbolicRep t => Symbolic t -> t) -> ContractInstanceKey state w s e p -> p -> Contract w s e ()

Map a ContractInstanceKey k to the Contract that is started when we start k in a given ModelState with a given semantics of SymTokens

perform

Arguments

:: HandleFun state
Function from ContractInstanceKey to ContractHandle

-> (forall t. HasSymbolicRep t => Symbolic t -> t)
Map from symbolic things (that may appear in actions or the state) to actual things that appear at runtime

-> ModelState state
The model state before peforming the action

-> Action state
The action to perform

-> SpecificationEmulatorTrace ()

While nextState models the behaviour of the actions, perform contains the code for running the actions in the emulator (see Plutus.Trace.Emulator). It gets access to the wallet contract handles, the current model state, and the action to be performed.

shrinkAction :: ModelState state -> Action state -> [Action state]

When a test involving random sequences of actions fails, the framework tries to find a minimal failing test case by shrinking the original failure. Action sequences are shrunk by removing individual actions, or by replacing an action by one of the (simpler) actions returned by shrinkAction.

See shrink for more information on shrinking.

monitoring

Arguments

:: (ModelState state, ModelState state)
Model state before and after the action

-> Action state
The action that was performed

-> Property

-> Property

The monitoring function allows you to collect statistics of your testing using QuickCheck functions like label, collect, classify, and tabulate. This function is called by propRunActions (and friends) for any actions in the given Actions.

Statistics on which actions are executed are always collected.

restricted :: Action state -> Bool

In some scenarios it's useful to have actions that are never generated randomly, but only used explicitly in DL scenario actions. To avoid these actions matching an anyAction when shrinking, they can be marked restricted.

type Actions state = Actions (WithInstances (WrappedState state))

data SomeContractInstanceKey state where

Constructors

Key :: (SchemaConstraints w s e, Typeable p) => ContractInstanceKey state w s e p -> SomeContractInstanceKey state

class HasSymbolics a where

Minimal complete definition

Nothing

Methods

getAllSymbolics :: a -> SymCollectionIndex

Model state

data ModelState state

contractState :: Lens (ModelState state1) (ModelState state2) state1 state2

currentSlot :: Getter (ModelState state) Slot

balanceChanges :: Getter (ModelState state) (Map (AddressInEra Era) SymValue)

balanceChange :: Wallet -> Getter (ModelState state) SymValue

minted :: Getter (ModelState state) SymValue

lockedValue :: ModelState s -> SymValue

symIsZero :: SymValue -> Bool

class Monad m => GetModelState (m :: Type -> Type)

Minimal complete definition

getModelState

viewModelState :: GetModelState m => Getting a (ModelState (StateType m)) a -> m a

viewContractState :: (Coercible (StateType m) state, GetModelState m) => Getting a state a -> m a

Get a component of the contract state using a lens.

type SymToken = Symbolic AssetId

class SymValueLike v where

Methods

toSymValue :: v -> SymValue

class TokenLike t where

Methods

symAssetIdValueOf :: SymValue -> t -> Quantity

symAssetIdValue :: t -> Quantity -> SymValue

invSymValue :: SymValue -> SymValue

toValue :: (SymToken -> AssetId) -> SymValue -> Value

The Spec monad

newtype Spec state a

Constructors

Spec

Fields

unSpec :: WriterT SymCreationIndex (ReaderT (Var SymIndex) (State (ModelState state))) a

wait :: forall state. ContractModel state => Integer -> Spec state ()

waitUntil :: forall state. ContractModel state => Slot -> Spec state ()

mint :: SymValueLike v => v -> Spec state ()

burn :: SymValueLike v => v -> Spec state ()

deposit :: SymValueLike v => Wallet -> v -> Spec state ()

Add tokens to the balanceChange of a wallet. The added tokens are subtracted from the lockedValue of tokens held by contracts.

withdraw :: SymValueLike v => Wallet -> v -> Spec state ()

Withdraw tokens from a wallet. The withdrawn tokens are added to the lockedValue of tokens held by contracts.

transfer

Arguments

:: SymValueLike v

=> Wallet
Transfer from this wallet

-> Wallet
to this wallet

-> v
this many tokens

-> Spec state ()

Transfer tokens between wallets, updating their balances.

createToken :: String -> Spec state SymToken

createTxOut :: String -> Spec state SymTxOut

createTxIn :: String -> Spec state SymTxIn

createSymbolic :: HasSymbolicRep t => String -> Spec state (Symbolic t)

assertSpec :: String -> Bool -> Spec state ()

type SpecificationEmulatorTrace = Eff (Writer [(String, SomethingWithSymbolicRep)] ': BaseEmulatorEffects)

registerToken :: String -> AssetId -> SpecificationEmulatorTrace ()

Register the real token corresponding to a symbolic token created in createToken.

registerTxOut :: String -> TxOut CtxUTxO Era -> SpecificationEmulatorTrace ()

Register the real TxOut corresponding to a symbolic TxOut created in createTxOut.

registerTxIn :: String -> TxIn -> SpecificationEmulatorTrace ()

Register the real TxIn corresponding to a symbolic TxIn created in createTxIn.

registerSymbolic :: HasSymbolicRep t => String -> t -> SpecificationEmulatorTrace ()

delay :: Integer -> SpecificationEmulatorTrace ()

`delay n` delays emulator execution by n slots

fromSlotNo :: SlotNo -> Slot

toSlotNo :: Slot -> SlotNo

Test scenarios

type DL state = DL (WithInstances (WrappedState state))

action :: ContractModel state => Action state -> DL state ()

waitUntilDL :: ContractModel state => Slot -> DL state ()

observe :: ContractModel state => String -> ((forall t. HasSymbolicRep t => Symbolic t -> t) -> ChainState -> Bool) -> DL state ()

anyAction :: DL state ()

anyActions :: Int -> DL state ()

anyActions_ :: DL state ()

forAllQ :: Quantifiable q => q -> DL s (Quantifies q)

elementsQ :: Eq a => [a] -> Quantification a

chooseQ :: (Arbitrary a, Random a, Ord a) => (a, a) -> Quantification a

Failures

assert :: String -> Bool -> DL s ()

assertModel :: String -> (ModelState state -> Bool) -> DL state ()

stopping :: DL state ()

weight :: Double -> DL state ()

getSize :: DL state Int

monitor :: (Property -> Property) -> DL state ()

Properties

Wallet contract handles

type SchemaConstraints w schema err = (Typeable w, Monoid w, ToJSON w, Typeable schema, ContractConstraints schema, Show err, Typeable err, ToJSON err, FromJSON err, ToJSON w, FromJSON w)

The constraints required on contract schemas and error types to enable calling contract endpoints (callEndpoint).

data StartContract state where

Constructors

StartContract :: (SchemaConstraints w s e, Typeable p) => ContractInstanceKey state w s e p -> p -> StartContract state

type HandleFun state = forall w schema err params. (Typeable w, Typeable schema, Typeable err, Typeable params) => ContractInstanceKey state w schema err params -> ContractHandle w schema err

A function returning the ContractHandle corresponding to a ContractInstanceKey. A HandleFun is provided to the perform function to enable calling contract endpoints with callEndpoint.

Model properties

propSanityCheckModel :: forall state. ContractModel state => Property

Sanity check a ContractModel. Ensures that wallet balances are not always unchanged.

propSanityCheckAssertions :: forall state. ContractModel state => Actions state -> Property

Sanity check a ContractModel. Ensures that all assertions in the property generation succeed.

propSanityCheckReactive :: forall state. (ContractModel state, Eq state) => Actions state -> Positive Integer -> Positive Integer -> Property

Sanity check a ContractModel. Ensures that nextReactiveState is idempotent.

Coverage checking options

data CoverageOptions

Options for controlling coverage checking requirements

checkCoverage tells you whether or not to run the coverage checks at all.

`endpointCoverageEq instance endpointName` tells us what percentage of tests are required to include a call to the endpoint endpointName in the contract at `instance`.

coverIndex is the coverage index obtained from the CompiledCodeIn of the validator.

defaultCoverageOptions :: CoverageOptions

Default coverage checking options are: * not to check coverage * set the requriements for every endpoint to 20% and * not to cover any source locations in the validator scripts.

endpointCoverageReq :: Lens' CoverageOptions (ContractInstanceTag -> String -> Double)

checkCoverage :: Lens' CoverageOptions Bool

coverageIndex :: Lens' CoverageOptions CoverageIndex

quickCheckWithCoverage :: Testable prop => Args -> CoverageOptions -> (CoverageOptions -> prop) -> IO CoverageReport

Run QuickCheck on a property that tracks coverage and print its coverage report.

quickCheckWithCoverageAndResult :: Testable prop => Args -> CoverageOptions -> (CoverageOptions -> prop) -> IO (CoverageReport, Result)

Emulator properties

propRunActions_

Arguments

:: ContractModel state

=> Actions state
The actions to run

-> Property

Run a Actions in the emulator and check that the model and the emulator agree on the final wallet balance changes. Equivalent to

propRunActions_ hs actions = propRunActions hs (const $ pure True) actions

propRunActions

Arguments

:: ContractModel state

=> (ModelState state -> TracePredicate)
Predicate to check at the end

-> Actions state
The actions to run

-> Property

Run a Actions in the emulator and check that the model and the emulator agree on the final wallet balance changes, and that the given TracePredicate holds at the end. Equivalent to:

propRunActions = propRunActionsWithOptions defaultCheckOptionsContractModel defaultCoverageOptions

propRunActionsWithOptions

Arguments

:: ContractModel state

=> CheckOptions
Emulator options

-> CoverageOptions
Coverage options

-> (ModelState state -> TracePredicate)
Predicate to check at the end

-> Actions state
The actions to run

-> Property

Run a Actions in the emulator and check that the model and the emulator agree on the final wallet balance changes, that no off-chain contract instance crashed, and that the given TracePredicate holds at the end. The predicate has access to the final model state.

The Actions argument can be generated by a forAllDL from a DL scenario, or using the Arbitrary instance for actions which generates random actions using arbitraryAction:

>>> quickCheck $ propRunActions_ handles
+++ OK, passed 100 tests
>>> quickCheck $ forAllDL dl $ propRunActions_ handles
+++ OK, passed 100 tests

The options argument can be used to configure the emulator--setting initial wallet balances, the maximum number of slots to run for, and the log level for the emulator trace printed on failing tests:

options :: Map Wallet Value -> Slot -> LogLevel -> CheckOptions
options dist slot logLevel =
    defaultCheckOptions & emulatorConfig . initialChainState .~ Left dist
                          & minLogLevel                        .~ logLevel

defaultCheckOptionsContractModel :: CheckOptions

Default check options that include a large amount of Ada in the initial distributions to avoid having to write ContractModels that keep track of balances.

checkThreatModel

Arguments

:: CheckableContractModel state

=> ThreatModel a

-> Actions (WithInstances state)
The actions to run

-> Property

Check a threat model on all transactions produced by the given actions.

checkThreatModelWithOptions

Arguments

:: CheckableContractModel state

=> CheckOptions
Emulator options

-> CoverageOptions
Coverage options

-> ThreatModel a

-> Actions (WithInstances state)
The actions to run

-> Property

Check a threat model on all transactions produced by the given actions.

DL properties

forAllDL :: (ContractModel state, Testable p) => DL state () -> (Actions state -> p) -> Property

Standard properties

type NoLockedFundsProof state = NoLockedFundsProof (WrappedState state)

pattern NoLockedFundsProof :: DL (WithInstances model) () -> (Wallet -> DL (WithInstances model) ()) -> (ModelState (WithInstances model) -> SymValue) -> (ModelState (WithInstances model) -> SymValue) -> NoLockedFundsProof model

nlfpMainStrategy :: NoLockedFundsProof model -> DL (WithInstances model) ()

Strategy to recover all funds from the contract in any reachable state.

nlfpWalletStrategy :: NoLockedFundsProof model -> Wallet -> DL (WithInstances model) ()

A strategy for each wallet to recover as much (or more) funds as the main strategy would give them in a given state, without the assistance of any other wallet.

nlfpOverhead :: NoLockedFundsProof model -> ModelState (WithInstances model) -> SymValue

An initial amount of overhead value that may be lost - e.g. setup fees for scripts that can't be recovered.

nlfpErrorMargin :: NoLockedFundsProof model -> ModelState (WithInstances model) -> SymValue

The total amount of margin for error in the value collected by the WalletStrategy compared to the MainStrategy. This is useful if your contract contains rounding code that makes the order of operations have a small but predictable effect on the value collected by different wallets.

defaultNLFP :: NoLockedFundsProof model

The default skeleton of a NoLockedFundsProof - doesn't permit any overhead or error margin.

checkNoLockedFundsProof :: CheckableContractModel model => NoLockedFundsProof model -> Property

Check a NoLockedFundsProof. Each test will generate an arbitrary sequence of actions (anyActions_) and ask the nlfpMainStrategy to recover all funds locked by the contract after performing those actions. This results in some distribution of the contract funds to the wallets, and the test then asks each nlfpWalletStrategy to show how to recover their allotment of funds without any assistance from the other wallets (assuming the main strategy did not execute). When executing wallet strategies, the off-chain instances for other wallets are killed and their private keys are deleted from the emulator state.

checkNoLockedFundsProofFast :: CheckableContractModel model => NoLockedFundsProof model -> Property

type NoLockedFundsProofLight state = NoLockedFundsProofLight (WrappedState state)

pattern NoLockedFundsProofLight :: DL (WithInstances model) () -> NoLockedFundsProofLight model

nlfplMainStrategy :: NoLockedFundsProofLight model -> DL (WithInstances model) ()

checkNoLockedFundsProofLight :: CheckableContractModel model => NoLockedFundsProofLight model -> Property

checkNoLockedFundsProofWithOptions :: CheckableContractModel model => CheckOptions -> NoLockedFundsProof model -> Property

checkNoLockedFundsProofFastWithOptions :: CheckableContractModel model => CheckOptions -> NoLockedFundsProof model -> Property

data Whitelist

A whitelist entry tells you what final log entry prefixes are acceptable for a given error

whitelistOk :: Whitelist -> Bool

Check that a whitelist does not accept any partial functions

mkWhitelist :: [Text] -> Whitelist

errorPrefixes :: Whitelist -> Set Text

defaultWhitelist :: Whitelist

checkErrorWhitelist :: ContractModel m => Whitelist -> Actions m -> Property

Check that running a contract model does not result in validation failures that are not accepted by the whitelist.

checkErrorWhitelistWithOptions :: forall m. ContractModel m => CheckOptions -> CoverageOptions -> Whitelist -> Actions m -> Property

Check that running a contract model does not result in validation failures that are not accepted by the whitelist.

newtype WithInstances s

Constructors

WithInstances

Fields

withoutInstances :: s

newtype WrappedState state

Constructors

WrapState

Fields

unwrapState :: state

checkDoubleSatisfaction :: forall m. ContractModel m => Actions m -> Property

Perform a light-weight check to find egregious double satisfaction vulnerabilities in contracts.

A counterexample to this property consists of three transactions. * The first transaction is a valid transaction from the trace generated by the contract model. * The second transaction, generated by redirecting a non-datum pubkey output from a non-signer to a signer in the first transaction, fails to validate. This demonstrates that funds can't simply be stolen. * The third transaction goes through and manages to steal funds by altering the first transaction. It is generated by adding another script input (with the same value as the non-signer non-stealable pubkey output) and adding a datum to the non-signer non-stealable pubkey output, and giving the extra value from the new script input to a signer.

checkDoubleSatisfactionWithOptions :: forall m. ContractModel m => CheckOptions -> CoverageOptions -> Actions m -> Property

Perform a light-weight check to find egregious double satisfaction vulnerabilities in contracts, with options.

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Minimal complete definition

from, to

Utils

fromAssetId :: AssetId -> AssetClass

toAssetId :: AssetClass -> AssetId

Orphan instances

HasSymbolics Wallet

Instance details

Methods

getAllSymbolics :: Wallet -> SymCollectionIndex

HasSymbolics Value

Instance details

Methods

getAllSymbolics :: Value -> SymCollectionIndex

HasSymbolics BuiltinByteString

Instance details

Methods

getAllSymbolics :: BuiltinByteString -> SymCollectionIndex

SymValueLike Ada

Instance details

Methods

toSymValue :: Ada -> SymValue

SymValueLike Value

Instance details

Methods

toSymValue :: Value -> SymValue

TokenLike AssetClass

Instance details

Methods

symAssetIdValueOf :: SymValue -> AssetClass -> Quantity

symAssetIdValue :: AssetClass -> Quantity -> SymValue

HasVariables BuiltinByteString

Instance details

Methods

getAllVariables :: BuiltinByteString -> Set (Any Var)